load render pipelines in parallel
