+++ Modules/TGraphics/Sources/TGraphics/TGPipelinesManager.swift
@@ -48,53 +48,23 @@ class TGPipelinesManager {
                 usingFunctionNamed: TGComputePipelineType.allCases[index].rawValue
             ) { [weak self] in
                 guard let self else { return }
-                self.computePipelineStates[index] = $0
-                self.dispatchGroup.leave()
+                computePipelineStates[index] = $0
+                dispatchGroup.leave()
             }
         }
         dispatchGroup.wait()
-       
-        // TODO: load as compute pipelines
-        renderPipelineStates[0] = makeRenderPipelieState(
-            withLabel: TGRenderPipelineType.drawTexture.label,
-            vertexFunction: TGRenderPipelineType.drawTexture.vertexFunction,
-            fragmentFunction: TGRenderPipelineType.drawTexture.fragmentFunction,
-            factory: { descriptor in
-                descriptor.colorAttachments[0].isBlendingEnabled = true
-                descriptor.colorAttachments[0].rgbBlendOperation = .add
-                descriptor.colorAttachments[0].alphaBlendOperation = .add
-                descriptor.colorAttachments[0].sourceRGBBlendFactor = .sourceAlpha
-                descriptor.colorAttachments[0].destinationRGBBlendFactor = .oneMinusSourceAlpha
-                descriptor.colorAttachments[0].sourceAlphaBlendFactor = .one
-                descriptor.colorAttachments[0].destinationAlphaBlendFactor = .oneMinusSourceAlpha
-            }
-        )
-        renderPipelineStates[1] = makeRenderPipelieState(
-            withLabel: TGRenderPipelineType.drawGrayScalePoints.label,
-            vertexFunction: TGRenderPipelineType.drawGrayScalePoints.vertexFunction,
-            fragmentFunction: TGRenderPipelineType.drawGrayScalePoints.fragmentFunction,
-            factory: { descriptor in
-                descriptor.colorAttachments[0].isBlendingEnabled = true
-                descriptor.colorAttachments[0].rgbBlendOperation = .max
-                descriptor.colorAttachments[0].alphaBlendOperation = .max
-                descriptor.colorAttachments[0].sourceRGBBlendFactor = .one
-                descriptor.colorAttachments[0].destinationRGBBlendFactor = .one
-                descriptor.colorAttachments[0].sourceAlphaBlendFactor = .one
-                descriptor.colorAttachments[0].destinationAlphaBlendFactor = .one
-                
-                let vertexDescriptor = MTLVertexDescriptor()
-                vertexDescriptor.attributes[0].format = .float2
-                
-                vertexDescriptor.attributes[1].format = .float
-                vertexDescriptor.attributes[1].offset = MemoryLayout<simd_float2>.stride
-                
-                vertexDescriptor
-                    .layouts[0]
-                    .stride = MemoryLayout<TGRenderablePoint>.stride
-                
-                descriptor.vertexDescriptor = vertexDescriptor
-            }
-        )
+        
+        for (index, type) in TGRenderPipelineType.allCases.enumerated() {
+            dispatchGroup.enter()
+            TGRenderPipelineStateFactory()
+                .makePipelineState(
+                    ofType: type) { [weak self] in
+                        guard let self else { return }
+                        renderPipelineStates[index] = $0
+                        dispatchGroup.leave()
+                    }
+        }
+        dispatchGroup.enter()
     }
     
     func computePipeline(ofType type: TGComputePipelineType) -> MTLComputePipelineState? {
@@ -127,29 +97,6 @@ class TGPipelinesManager {
                 debugPrint(error!)
                 fatalError()
             }
-        }
-    }
-    
-    private func makeRenderPipelieState(
-        withLabel label: String,
-        vertexFunction: String,
-        fragmentFunction: String,
-        factory: (MTLRenderPipelineDescriptor) -> Void
-    ) -> MTLRenderPipelineState {
-        let library = TGDevice.defaultLibrary
-        let descriptor = MTLRenderPipelineDescriptor()
-        descriptor.vertexFunction = library.makeFunction(name: vertexFunction)
-        descriptor.fragmentFunction = library.makeFunction(name: fragmentFunction)
-        descriptor.colorAttachments[0].pixelFormat = .rgba8Unorm
-        descriptor.label = label
-        
-        factory(descriptor)
-        
-        do {
-            return try TGDevice.device.makeRenderPipelineState(descriptor: descriptor)
-        } catch {
-            debugPrint(error)
-            fatalError()
         }
     }
 }
